chore: 版本统一 3.0.0，WebUI 模块化并完善卡片渲染
- metadata/main 与 WebUI 缓存参数统一为 3.0.0
- app.js 拆分为 state/data/ui/nav/pages 等模块
- 设置 schema 由 webui_settings_schema 同源导出
- 卡片/公式渲染与字体探测增强
- 同步 README / CHANGELOG 文案
