+++ README.md
@@ -76,8 +76,8 @@ hapi codex    # Open Codex
 - **文件双向传输**: 支持利用astrbot进行小文件的下载、上传，方便查看日志或传递配置
 - **兼容QQ、微信的官方bot**: 无法主动推送消息时将消息fallback伪装为被动回复，兼容QQ官方bot、微信clawbot
 - **智能审批机制**: 支持戳一戳快速批准、忙时自动托管、超时提醒，灵活应对不同场景
-- **Web 管理面板**: 在 AstrBot 插件详情页可视化查看连接与 session、改推送/绑定/权限、保存插件配置（与官方设置页同源）
-- **推送呈现（可选）**: 结构卡与 **Agent 对话** 同一套卡片管线；支持自定义 CSS；Pillow=低延迟快路径，Playwright=完整 CSS；中文字体用 card_font_path / assets/fonts / 系统 CJK（不自动下载），未装引擎或无字体时回退文本
+- **交互友好的 Web 管理面板**: 在 AstrBot 插件详情页可视化查看连接与 session、改推送/绑定/权限、保存插件配置（与官方设置页同源）
+- **低延迟、稳定的推送服务与结果渲染**: 相较于常规内网穿透方案更加稳定，可在webui选择下载依赖渲染结果为图片(使用pillow而非笨重的playwright)
 
 
 ## 💡 实际应用场景
@@ -359,7 +359,7 @@ astrbot_plugin_hapi_connector/
 - ✅ 支持文件上传与下载逻辑
 - ✅ 支持多用户独立会话状态，通知相互隔离
 - ✅ 通过 AstrBot 自然语言触发指令
-- [ ] 支持将 Markdown 文字、AI编辑等影响观感长上下文渲染为图片（依赖库独立，可选下载）
+- ✅ 支持将 Markdown 文字、AI编辑等影响观感长上下文渲染为图片（依赖库独立，可选下载）
 
 ---
