+++ pages/console/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>HAPI 连接器 · 管理面板</title>
     <!-- 纯静态本地资源，符合 plugin-pages：禁止外链 CDN / 字体 / 脚本 -->
-    <link rel="stylesheet" href="./style.css?v=3.1.42" />
+    <link rel="stylesheet" href="./style.css?v=3.0.0" />
   </head>
   <body>
     <div class="app-shell" id="app">
@@ -132,6 +132,6 @@
       </div>
     </dialog>
 
-    <script type="module" src="./app.js?v=3.1.42"></script>
+    <script type="module" src="./app.js?v=3.0.0"></script>
   </body>
 </html>
